Created Arduino_Mega_ISR_Registry class.
Arduino_Mega_ISR_Registry provides an implementation of the interrupt
signals from avr/interrupt.h, and a class encapsulation which allows
a callback to be registered for an ISR.

It is often necessary to attach a piece of code to an interrupt during
run-time because the Arduino environment will link every object file in
a given library. In order for mutually exclusive implementations of
ISR code to exist in a library, they cannot be declaread as ISR()'s because
the linker will find multiple implementations. Instead, a callback must
be attached to a function during initialization.

Presently two ISRs are implemented, TIMER2_OVF and TIMER4_CAPT. More
will be added as they are required.